requirements.txt
Flask==1.1.2
Jinja2==2.11.2
MarkupSafe==1.1.1
itsdangerous==1.1.0
gunicorn==20.0.4
line-bot-sdk==1.17.0
requests==2.25.0
urllib3<2.0
click==7.1.2
certifi==2020.11.8
chardet==3.0.4
future==0.18.2
idna==2.10
Werkzeug==1.0.1
python-dotenv
pymongo
yfinance
schedule
twstock